requirements.txt
anthropic>=0.40.0
langgraph>=0.2.0
langchain-core>=0.3.0
tavily-python>=0.5.0
openai>=1.50.0
streamlit>=1.39.0
pydantic-settings>=2.6.0
pydantic>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
pytest>=8.3.0
pytest-asyncio>=0.24.0
